Duotone: moving components away from duotone hook file
