+++ packages/block-editor/src/components/block-preview/auto.js
@@ -12,7 +12,7 @@ import { useMemo } from '@wordpress/element';
 import BlockList from '../block-list';
 import Iframe from '../iframe';
 import EditorStyles from '../editor-styles';
-import { PresetDuotoneFilter } from '../../hooks/duotone';
+import { PresetDuotoneFilter } from '../../components/duotone/index';
 import { store } from '../../store';
 
 // This is used to avoid rendering the block list if the sizes change.
