+++ packages/block-editor/src/components/duotone/index.js
@@ -1,0 +1,2 @@
+export * from './utils';
+export * from './components';
